requirements.txt
flask>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scikit-surprise>=1.1.3
openai>=1.0.0
python-dotenv>=1.0.0
scipy>=1.11.0
